+++ 2D/2SF.cpp
@@ -28,7 +28,7 @@ public:
 
 class SF : public Bitmap<Point> {
 public:
-    SF(const Hub &H, int n_, double a_)
+    SF(const Hub &H, size_t n_, double a_)
         : Bitmap<Point>(H.title, {2 * n_, 2 * n_ + 1}), n(n_), a(a_), root({2 * n - 1, 2 * (n / 2)}), start({1, 2 * (n / 4)}) {
         ps = {a * a, a * a, 1, 1};
         for (auto &p : ps) p /= 2 * (1 + a * a);
