+++ 2D/DAC.cpp
@@ -7,7 +7,7 @@ using namespace std;
 
 int main(int argc, char **argv) {
     Hub    H("Divide-and-color model", argc, argv, "n=300,p=.3,q=.4");
-    int    n = H['n'];
+    size_t n = H['n'];
     double p = H['p'], q = H['q'];
 
     // Connectivity (&1 -> to the right, &2 -> downwards)
