+++ 2D/Ising-bis.cpp
@@ -12,7 +12,7 @@ namespace vb {
 
 class Ising : public Bitmap<int> {
 public:
-    Ising(const Hub &H, int n, double beta, int con) : Bitmap<int>(H.title, {n, n}), con(con), p(2 * con + 1) {
+    Ising(const Hub &H, size_t n, double beta, int con) : Bitmap<int>(H.title, {n, n}), con(con), p(2 * con + 1) {
         for (auto z : coo_range(size)) put(z, prng.bernoulli(.5) ? 1 : -1);
         if (con == 6) {
             int m = 2 * n / 3;
