+++ 2D/Ising-cftp.cpp
@@ -15,7 +15,8 @@ namespace vb {
 
 class IsingCFTP : public Bitmap<int> {
 public:
-    explicit IsingCFTP(const Hub &H) : Bitmap<int>(H.title, {H['n'], H['n']}), b(H['b']), d(0), s(H['s']), status(H.title, {w(), h()}) {
+    explicit IsingCFTP(const Hub &H)
+        : Bitmap<int>(H.title, {H['n'], H['n']}), b(H['b']), d(0), s(H['s']), status(H.title, {size_t(w()), size_t(h())}) {
         for (int i = 0; i < w(); ++i)
             for (int j = 0; j < h(); ++j) put({i, j}, 1);
         snap();
