+++ 2D/LERW.cpp
@@ -11,7 +11,7 @@ namespace vb {
 
 class LERW : private Bitmap<int> {
 public:
-    explicit LERW(const Hub &H) : Bitmap<int>(H.title, {2 * int(H['n']), 2 * int(H['n'])}) {
+    explicit LERW(const Hub &H) : Bitmap<int>(H.title, {2 * size_t(H['n']), 2 * size_t(H['n'])}) {
         coo z{w() / 2, h() / 2};
         while (contains(z)) {
             int d = prng() & 3u;
