+++ 2D/PercSEP.cpp
@@ -15,7 +15,7 @@ namespace vb {
 
 class PercSEP : public Bitmap<int> {
 public:
-    explicit PercSEP(const Hub &H) : Bitmap<int>(H.title, {2 * int(H['n']), H['n']}), flow({0, 0}), d(H['d']), tasym(H['t']) {
+    explicit PercSEP(const Hub &H) : Bitmap<int>(H.title, {2 * size_t(H['n']), H['n']}), flow({0, 0}), d(H['d']), tasym(H['t']) {
         for (int i = 0; i < w(); ++i)
             for (int j = 0; j < h(); ++j)
                 if (prng.bernoulli(H['p'])) put({i, j}, prng.bernoulli(H['l']) ? 2 : 1);
