+++ 2D/Snake2.cpp
@@ -18,7 +18,8 @@ namespace vb {
 
 class Snake : public Bitmap<site> {
 public:
-    Snake(const Hub &H, int n, double l, bool hex) : Bitmap(H.title, {4 * n + 1, 2 * n + 1}), lambda(l), path(1, {2 * n, 0}), hex(hex) {
+    Snake(const Hub &H, size_t n, double l, bool hex)
+        : Bitmap(H.title, {4 * n + 1, 2 * n + 1}), lambda(l), path(1, {2 * int(n), 0}), hex(hex) {
         for (int x = 0; x < w(); ++x) put({x, 0}, VERTEX);
         if (double a = H['a']; a != 0) triangle(a);
         if (!H['v']) show();
